[FIX] : divide date class
