+++ app/src/main/java/com/dosopt/naverpay/domain/place/PlaceBrandList.kt
@@ -1,0 +1,6 @@
+package com.dosopt.naverpay.domain.place
+
+data class BrandList(
+    val id: Long,
+    val logo_img_url: String
+)
