+++ app/src/main/java/com/dosopt/naverpay/domain/place/PlaceOnsitePaymentList.kt
@@ -1,0 +1,7 @@
+package com.dosopt.naverpay.domain.place
+
+data class OnsitepaymentList(
+    val id: Long,
+    val logo_img_url: String,
+    val name: String
+)
