+++ app/src/main/java/com/dosopt/naverpay/domain/place/PlaceUser.kt
@@ -1,0 +1,5 @@
+package com.dosopt.naverpay.domain.place
+
+data class User(
+    val user_name: String = ""
+)
